메인 파일 삭제 및 JavaApplication 클래스 생성 및 main문 구현
