--- src/main/java/com/sprint/mission/Main.java
@@ -1,10 +1,0 @@
-package com.sprint.mission;
-
-//TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
-// click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
-public class Main {
-    public static void main(String[] args) {
-
-        }
-    }
-}
